+++ src/Containers/BoardPage.tsx
@@ -5,6 +5,7 @@ import BoardContent from '../components/BoardContent';
 import { makeStyles } from '@material-ui/styles';
 import { Theme } from '@material-ui/core/styles';
 import firebase from 'firebase';
+import SnackbarContent from '@material-ui/core/SnackbarContent';
 
 /**
  * Returns a page where it displays a kanban style board to users, the sidebar, and the header component
@@ -15,15 +16,27 @@ const useStyles = makeStyles((theme: Theme) => ({
     marginTop: theme.spacing(3),
     flexGrow: 1,
   },
+  padding: {
+    padding: theme.spacing(2),
+  },
+  snackbar: {
+    backgroundColor: '#428bca',
+  },
 }));
 
 const Board: React.FC = (props: any) => {
   const classes = useStyles();
+  const [showWarning, setShowWarning] = React.useState<boolean>(false); //state to display if profile name has been set. Else show a warning to prompt profile update
+
   //on component mount, check if user is logged in, else signs out the user.
   React.useEffect(() => {
     firebase.auth().onAuthStateChanged(function(user) {
       if (user) {
-        // User is signed in. Do nothing
+        if (user.displayName === null) {
+          setShowWarning(true); //display set profile message
+        } else {
+          setShowWarning(false);
+        }
       } else {
         handleLogout(); //logout user
       }
@@ -48,6 +61,14 @@ const Board: React.FC = (props: any) => {
       <SideBar />
       <div className={classes.root}>
         <BoardHeader />
+        {showWarning && (
+          <div className={classes.padding}>
+            <SnackbarContent
+              message={'It appears you do not have a display name! Update your profile to create user stories.'}
+              className={classes.snackbar}
+            />
+          </div>
+        )}
         <BoardContent />
       </div>
     </div>